# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: PayPass API
  description: |
    PayPass is a comprehensive payment platform providing secure, scalable payment processing solutions.
    
    ## Features
    - Payment Processing
    - Wallet Management  
    - User Authentication
    - Enterprise Billing
    - Fraud Detection
    - Compliance Management
    
    ## Authentication
    PayPass API uses API key authentication. Include your API key in the Authorization header:
    ```
    Authorization: Bearer YOUR_API_KEY
    ```
    
    ## Rate Limiting
    API requests are rate limited based on your subscription plan:
    - Basic: 100 requests/hour
    - Premium: 1000 requests/hour  
    - Enterprise: 10000 requests/hour
    
    ## Webhooks
    PayPass can send webhook notifications for various events. Configure webhooks in your dashboard.
    
  version: 2.1.0
  contact:
    name: PayPass API Support
    email: api-support@paypass.com
    url: https://docs.paypass.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: https://paypass.com/terms
servers:
  - url: https://api.paypass.com/v1
    description: Production server
  - url: https://api-staging.paypass.com/v1
    description: Staging server
  - url: https://api-dev.paypass.com/v1
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /auth/refresh:
    post:
      tags:
        - Authentication
      summary: Refresh access token
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                refresh_token:
                  type: string
      responses:
        '200':
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'

  # Enterprise Endpoints
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: User ID
        email:
          type: string
          format: email
          description: Email address
        first_name:
          type: string
          description: First name
        last_name:
          type: string
          description: Last name
        phone:
          type: string
          description: Phone number
        country:
          type: string
          description: Country
        status:
          type: string
          enum: [active, inactive, suspended]
          description: User status
        created_at:
          type: string
          format: date-time
          description: Creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp

    # Wallet Schemas
...
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
          description: JWT access token
        refresh_token:
          type: string
          description: Refresh token
        token_type:
          type: string
          example: Bearer
          description: Token type
        expires_in:
          type: integer
          description: Token expiry in seconds
        user:
          $ref: '#/components/schemas/User'

    # Enterprise Schemas
...